+++ test/assertions_test.js
@@ -2024,6 +2024,90 @@ describe("Must.prototype.own", function() {
   })
 })
 
+describe("Must.prototype.keys", function() {
+  it("must pass given an object with expected keys", function() {
+    assertPass(function() { Must({a: 1, b: 2}).have.keys(["a", "b"]) })
+  })
+
+  it("must pass given an object with zero keys", function() {
+    assertPass(function() { Must({}).have.keys([]) })
+  })
+
+  it("must fail given zero and non-zero number of keys", function() {
+    assertFail(function() { Must({}).have.keys(["a"]) })
+    assertFail(function() { Must({a: 1}).have.keys([]) })
+  })
+
+  it("must fail given a different amount of keys", function() {
+    assertFail(function() { Must({a: 1}).have.keys(["a", "b"]) })
+    assertFail(function() { Must({a: 1, b: 2}).have.keys(["a"]) })
+  })
+
+  describe("given an inherited object", function() {
+    it("must pass given an object with some expected keys inherited",
+      function() {
+      var obj = Object.create({a: 1}, {b: {value: 2, enumerable: true}})
+      assertPass(function() { Must(obj).have.keys(["a", "b"]) })
+    })
+
+    it("must pass given an object with all expected keys inherited",
+      function() {
+      var obj = Object.create({a: 1, b: 2})
+      assertPass(function() { Must(obj).have.keys(["a", "b"]) })
+    })
+
+    it("must fail given an object when expecting zero keys", function() {
+      assertFail(function() { Must(Object.create({a: 1})).have.keys([]) })
+    })
+
+    it("must fail given a different amount of keys", function() {
+      assertFail(function() {
+        Must(Object.create({a: 1})).have.keys(["a", "b"]) 
+      })
+      assertFail(function() {
+        Must(Object.create({a: 1, b: 2})).have.keys(["a"]) 
+      })
+    })
+  })
+
+  it("must fail gracefully if null", function() {
+    assertFail(function() { Must(null).have.keys(["love"]) })
+  })
+
+  it("must fail gracefully if undefined", function() {
+    assertFail(function() { Must(undefined).have.keys(["love"]) })
+  })
+
+  it("must fail gracefully if non-object", function() {
+    assertFail(function() { Must(true).have.keys(["love"]) })
+  })
+
+  it("must be bound", function() {
+    var obj = {a: 1, b: 2}
+    assertPass(function() { obj.must.have.keys.call(null, ["a", "b"]) })
+  })
+
+  mustThrowAssertionError(function() { ({a: 1}).must.have.keys(["a", "b"]) }, {
+    actual: {a: 1},
+    expected: ["a", "b"],
+    message: "{\"a\":1} must have keys [\"a\",\"b\"]"
+  })
+
+  describe(".not", function() {
+    function not() { ({a: 1, b: 2}).must.not.have.keys(["a", "b"]) }
+
+    it("must invert the assertion", function() {
+      assertFail(not)
+    })
+
+    mustThrowAssertionError(not, {
+      actual: {a: 1, b: 2},
+      expected: ["a", "b"],
+      message: "{\"a\":1,\"b\":2} must not have keys [\"a\",\"b\"]"
+    })
+  })
+})
+
 function mustPassEnumerable(name, truthy) {
   var pass = truthy ? "pass" : "fail"
   var fail = truthy ? "fail" : "pass"