+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Unreleased
-- Fails gracefully if property matchers (`property`, `enumerable` etc.) are used
-  on non-objects.
+- Fails gracefully if property matchers ([`property`][],
+  [`enumerable`][] etc.) are used on non-objects.
+- Adds the [`keys`][] matcher to test if an object has all
+  the expected keys.  
+  Takes inherited keys into account just like the [`empty`][] and [`property`][]
+  matchers.
+
+[`empty`]: https://github.com/moll/js-must/blob/master/doc/API.md#Must.prototype.empty
+[`enumerable`]: https://github.com/moll/js-must/blob/master/doc/API.md#Must.prototype.enumerable
+[`property`]: https://github.com/moll/js-must/blob/master/doc/API.md#Must.prototype.property
+[`keys`]: https://github.com/moll/js-must/blob/master/doc/API.md#Must.prototype.keys
 
 ## 0.1.338 (Sep 26, 2013)
 - Fixes Must.js's own tests on Windows thanks to [Bart van der Schoor][bartvds].